chore: make cargo clippy -D warnings green across workspace
Closes gap P1 #12. .github/workflows/ci.yml already had the gate
wired as cargo clippy --workspace --exclude onebit-hip --exclude
onebit-mlx --release --all-targets --no-deps -- -D warnings
-A clippy::uninlined-format-args. The gate was effectively a no-op
because clippy had 49 unique issues across 22 distinct errors
(73 raw with lib+test target duplication).

This commit flips the gate on by fixing every warning.

Categories:
 - 13x doc_overindented_list_items + doc_lazy_continuation
   (rustdoc list-block formatting in 1bit-helm/app, medusa/heads,
    medusa/loader, 1bit-watch-discord)
 - 12x needless_range_loop + unnecessary_cast
   (iter/enumerate rewrites in router, medusa, bitnet-to-tq2,
    gguf-to-h1b)
 - 8x identity_op + unusual_byte_groupings
   (shape-annotation identity ops 1 * K/2 kept as a clarity aid,
    date-shaped seeds 0x2026_04_20 kept for provenance, both with
    justified #[allow])

Flagged as follow-up gap-p2 with minimal #[allow] + TODO comments:
 - router::generate_blocking 8-arg signature needs GenCtx struct
 - h1b-sherry::build_layer_owned 8-arg signature needs LayerDims
 - MedusaState::Enabled payload variant needs boxing
 - CheckOutcome::Available 264-byte Artifact needs boxing

No logic changes. 455 lib tests green (was 436+, went up as
other agent work in the same window added registry and helm tests).

+++ tools/h1b-sherry/src/sherry_pack.rs
@@ -136,7 +136,7 @@ pub fn pack_sherry_row(ternary: &[i8], packed: &mut [u8], cols: usize) -> PackRo
         let bit_pos = 5 * g;
         let byte_idx = bit_pos >> 3;
         let shift = (bit_pos & 7) as u32;
-        packed[byte_idx] |= (code as u8).checked_shl(shift).unwrap_or(0);
+        packed[byte_idx] |= code.checked_shl(shift).unwrap_or(0);
         // Overflow into next byte when a 5-bit code straddles a byte boundary.
         if shift + 5 > 8 {
             packed[byte_idx + 1] |= code >> (8 - shift);
